pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Use automatic package discovery and include only real Python packages
[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "db*", "ml*", "scripts*"]

[project]
name = "ai-risk-monitor"
version = "0.1.0"
description = "FastAPI gateway for AI risk & compliance monitoring"
requires-python = ">=3.10"
authors = [{name = "Your Name"}]
readme = "README.md"
license = {text = "Apache-2.0"}
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "prometheus-client>=0.20.0",
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "anyio>=4.4.0",
  "ruff>=0.6.5",
  "chromadb>=0.5.4",
  "sqlalchemy>=2.0.32",
  "sentence-transformers==3.0.1",
  "openai>=1.51.2",
  "pymupdf>=1.24.9",
  "mlflow>=2.15.0",
  "scikit-learn>=1.5.0",
  "pandas>=2.2.0",
]


[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I"]
lint.ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]
"app/schemas/research.py" = ["E402"]
"tests/test_memory_endpoints.py" = ["E402"]
"tests/test_memory_long.py" = ["E402"]
"tests/test_memory_short.py" = ["E402"]
